Unify gfe2::RecordingHeadersHandler and spdy::TestHeadersHandler to spdy::RecordingHeadersHandler.
This CL extracts spdy::TestHeadersHandler into its own file and renames to
spdy::RecordingHeadersHandler. Then gfe2::RecordingHeadersHandler is nixed in
favor of spdy::RecordingHeadersHandler. This is a clean-up in its own right.

This CL is also a prefactor for forking http2_trace_logging.{h,cc} to
shinkansen_trace_logging.{h,cc}. It aims to reduce the content in
http2_trace_logging.{h,cc}, to reduce the content to fork.

PiperOrigin-RevId: 341693359
Change-Id: I00740b20a4e95dca9f612fe8a4dc155814a7b84f