Implement a flag-protected version of the QuicIntervalSet performance improvements.
I tried to make it easy to review:  The diff between the first two snapshots the code that simply duplicates the old code into the new code, putting the two versions of the code into two different naespaces (newquic and oldquic).  Then the diff from the second snapshot to the present is all the changes to the oldcode to make it work as the newcode

(Originally this appeared as a set of changes: cl/337932377 cl/337932365 cl/337783824 cl/337932380 cl/337932384 cl/337932371 cl/337932361.)

Protected by FLAGS_quic_restart_flag_quic_startup_faster_interval_set.

PiperOrigin-RevId: 343903036
Change-Id: I9f9aaad3d8a86be0e1426e06e5dc3f6e7d2c5476