Prevent buffer overflow in QuicSimpleServerSession::HandleRstOnValidNonexistentStream
This fix is based on code from <liujiyong4@gmail.com>, thanks to them for reporting the issue and proposing a fix.

The issue here was that this code was using DCHECKs to validate input from the network instead of failing gracefully. QuicSimpleServerSession is not used in production so this code does not require flag protection.

PiperOrigin-RevId: 344319749
Change-Id: Ic0f14412ea5e2b398b48cb3d2333c2e1e01d6d35