Move enforcement of single RETRY from QuicConnection to QuicFramer
This change is a no-op. Before this change, we would first validate the retry integrity tag, and then check whether we had already received a RETRY. The issue with that was that it would QUIC_DLOG(ERROR) when failing to validate. With this change, we drop that second RETRY without validating it, removing the log. This log was only in debug builds so this behavior change does not impact production. This CL also adds a new flag to the toy client which helped debug this issue.

PiperOrigin-RevId: 347645803
Change-Id: Iea5943c115f1267ee906f476bca17a184dedee4d