Unify Keystore testing tools across GFE
This CL eliminates all but one of the current N ways we configure a local Keystore instance in GFE (and Mentat) tests.  This results in the unwinding of a large amount of technical debt accreted over the last decade.

Note that there are associated changes to test certificates which can be found in the diffbase.

One notable change is the deletion of a test which checks that GFE fetches the TCP Fast Open test key.  This check is hard to implement with the local keystore tool, and I don't think the test case is useful (I also wrote it originally), so I've just deleted it.

Protected by Not protected; test changes only.

PiperOrigin-RevId: 354592197
Change-Id: Icb5eeb139a4ad8b122c3a31a6018a5d902b9bb6c