Fix use after free in MasqueServerBackend
This fixes some crashes in CloseBackendResponseStream that
were due to the fact that QuicConnection gets destructed before
the QuicSimpleServerBackend::RequestHandler is. Since we only
need to hold this memory to ensure the request handler is tracked,
we instead keep a vector of them per connection and clear the
entire vector when the connection goes away.

This code isn't used in production.

PiperOrigin-RevId: 360225184
Change-Id: I661ca92d2039adcc095eb4cd1a8ec6fa3ad9682b