Prevent QuicSimpleServerStream from sending two responses
If the request headers are invalid, QuicSimpleServerStream sends an error response. If the request was a form of CONNECT, this could lead us to incorrectly send two responses. This CL adds a bool to ensure we don't do that, and some DCHECKs for future sanity. This issue was found during the IETF 110 hackathon. This code is in test-only tools which are not used in production.

PiperOrigin-RevId: 360948797
Change-Id: Ia9dd74973efaf0cbd6f565db30ed46af31f4db1b