Document that HTTP/2 decoder does not segment frame payloads or fields.
I verified the added statements by code inspection.

The motivation is https://crrev.com/c/2728175: I'm adding a class which decodes
a stream of HTTP/2 frames that is guaranteed to be received in a single chunck.
It would allow me to simplify the implementation if I could rely on
ExtensionVisitorInterface::OnFramePayload() being called once with the entire
payload.
PiperOrigin-RevId: 361816447
Change-Id: I2c8621ceb52c2e17fdb84da91a97fb72e5c4450e