Modify old-style derived GFE_BUG macros to have IDs.
Usage of the macros was replaced by V2 set of macros. Context: go/gfe-bug-improvements. Now that everything was migrated to V2, we are updating V1 macros and will updated everything back to drop the V2 suffix.

GFE_BUG macros were changed in cl/362922435
SPDY_BUG macros were changed in cl/363160900, but one was missed. This CL corrects it.

PiperOrigin-RevId: 363240499
Change-Id: I77b4932e8e47bfef2f34000c0f5984e0a3b6ce45