Add quiche_string_view.h so that absl::string_view can be used in Chromium.
absl::string_view is used extensively in QUICHE, and it is used in Chromium
//net code only where it is absolutely necessary to interface with QUICHE.
However, presubmit rules do not allow the header to be included in Chromium,
except for QUICHE files that live in //net/third_party/quiche/src and form a
separate compilation unit.  Therefore we need a platform file to include
absl/string_view.h.  This file does not need to have a corresponding impl.

There was such file at quiche/common/platform/api/quiche_string_piece.h, which
has been removed at
https://quiche.googlesource.com/quiche/+/ccf4ebe76fc5c3a176f6128f201801046a686a29
This CL reinstates the include portion of that file, with a more appropriate
name.

PiperOrigin-RevId: 363913393
Change-Id: I86c445d3f6f2d8aab15f585590a55b4f921af1ed