Make HpackEntry::insertion_index_ start from 0 for dynamic entries.
Adding a constant offset of 61 does not add any value.  The calculation of
the HPACK index (which is 61 for the most recently added entry and growing for
older entries) involves a subtraction in GetByName() and GetByNameAndValue(),
and this offset cancels between total_insertion_ and entry->InsertionIndex(),
that's why that formula does not change with this CL.

Also add a couple of CHECKs involving InsertionIndex() to validate the formula
that will be used in the next CL which removes InsertionIndex().

PiperOrigin-RevId: 363924862
Change-Id: Ieea5a0a906229427f4af2c14f57074cf1e299b63