Resize HttpDecoder::buffer_ when buffering data.
Before this change, HttpDecoder::buffer_ was always empty.  A
std::string::reserve() call made sure its storage was large enough, and that
memory was address directly, always beyond the size of the string object.

After this CL, the size of buffer_ reflects the amount of buffered bytes, and
memory beyond its size is never addressed.  The reserve() call is preserved to
avoid reallocation and copy as the buffer grows.

PiperOrigin-RevId: 365076891
Change-Id: If58949fd10fe1741416dab4691faa5c3e728ad8e