Prepare QpackHeaderTable::InsertEntry() for non-stable container.
My plan is to split QpackHeaderTable: the decoder does not need
FindHeaderField(), or static_index_, static_name_index, dynamic_index_,
dynamic_index_name_ methods, therefore it does not need iterator stability for
the underlying container, but it needs random access.  QuicCircularDeque would
be a good choice.  This CL prepares InsertEntry() for that.

Since copy is done in a separate step before insertion, now it is safe to evict
before insertion, potentially saving on a little bit of memory.

PiperOrigin-RevId: 365609653
Change-Id: I55b6a95168d7a6bca32b9f35c2bf8d3c08472635