Use circular buffer instead of std::deque for HPACK dynamic table in HTTP/2 decoder.
Http2Deque is an alias for QuicheDequeImpl which is an alias for std::deque both
in internal code and in Chromium.  QuicCircularDeque is more memory efficient,
especially if there are only few entries, and requires fewer allocations as new
entries are added if old entries are also being removed, which is the case for
HPACK where the dynamic table size is bounded.

PiperOrigin-RevId: 366295336
Change-Id: I9457344c632ae26def506592d64a3d233cb380c1