Use QuicCircularDeque instead of QuicQueue.
QuicQueue is std::queue which is backed by std::deque in internal code.  It is
base::queue which a typedef for std::queue backed by base::circular_deque in
Chromium.

This changes the internal implementation from a container that has iterator
stability from one that does not (elements may move when a new one is added).
It is safe to do so because (1) by code inspection I established that nothing is
done with these two containers that would rely on iterator stability (like
taking address of front() then pushing a new element), and (2) this has already
been the case in Chromium for a while and there are no reported crashes or MSAN
failures.

My motivation is to remove as many things from QUICHE platform as possible.
Also note that QuicCircularDeque typically uses less memory than std::deque,
especially for small containers.

PiperOrigin-RevId: 366309029
Change-Id: I0f8f8185feee071b1043b7a60e6456444a281fab