In WebTransport, make FIN be a return of Read() method rather than a visitor call.
Currently, we notify the reader about FIN being read by calling Visitor::OnFinRead inside Read().  This has a problem that FIN can be processed *before* the final chunk of data in the stream (which would be processed after Read() returns).

As far as I can tell, the only reason this code works as-is is mostly a series of coincidences, and as soon as I tried to exercise it more carefully in one of WebTransport over HTTP/3 tests, it failed.

PiperOrigin-RevId: 366824443
Change-Id: Ie803f2ed206333eba683065f6251216a9c549fd7