Improve QpackHeaderTableTests.
Test for return value of SetMaximumDynamicTableCapacity() and
SetDynamicTableCapacity() when setting up QpackHeaderTableTests (must be done in
SetUp() override instead of constructor).

Use QpackHeaderTableTest::table_ in
QpackEncoderHeaderTableTest.MaxInsertSizeWithoutEvictingGivenEntry and
QpackEncoderHeaderTableTest.DrainingIndex.  Turns out the maximum dynamic table
capacity set in the test fixture is just fine for
MaxInsertSizeWithoutEvictingGivenEntry test.  (It used to use a local
QpackEncoderHeaderTable because SetMaximumDynamicTableCapacity() may only be
called once.)

PiperOrigin-RevId: 368414192
Change-Id: Ie40aa5f45cbf1675f876bc146d6fc733bb9cacd5