Make sure that a low QuicPacketCreator soft limit does not cause issues
This issue was causing a DCHECK failure in Envoy QUIC integration tests. In this test, Envoy would receive a QUIC packet and queue up some 1-RTT stream data which would trigger coalescing and set the soft max packet length. Then the code would try to send an initial ack. The issue there is that when switching the encryption level to initial, the header size would increase - and then when trying to add a frame we would trigger a DCHECK in BytesFree because the soft max packet length is lower than the header size. After performing code analysis, it is clear that any caller of BytesFree() properly handles the case where BytesFree() returns 0, so this DCHECK can be safely removed as it provides no value. This CL adds a regression test that would trigger the DCHECK if it were still present.

PiperOrigin-RevId: 368828653
Change-Id: I300326dd4e068569f4cec2690002cd4352d2afc2