Mark gfe2_reloadable_flag_quic_and_tls_allow_sni_without_dots verified
Submitting without explicit verification.  I spent a couple of hours trying to set up a staging L7XLB instance to test the functionality, but got blocked in a bunch of different ways, and it doesn't seem worth another several hours of frustration.  There are no flag counters because the actual mechanism is in httpsconnection code.  But it's also quite straightforward and has good test coverage.

PiperOrigin-RevId: 370565550
Change-Id: Iba625c17e04180f591e962c0c5afcd549fab584e