Defer TUN device creation until after the quic connection has been established.
IOS XR's XR container takes a few seconds for networking state to be sync'd from the host -- during which time any manipulations to the linux contaier's host networking stack breaks the synchronization process.

This has the nice side-benefit of re-initializing the TUN device each time a new quic connection is established, ensuring the local interface is in a good state on each reconnect. Previously, if the interface was down'd or deleted externally, the client would just fail continuously without attempting to recover.

PiperOrigin-RevId: 372620094