Automated g4 rollback of changelist 372620094.
*** Reason for rollback ***

Breaking Endor because we bring down the TUN Device when tearing down the bonnet in the init container.

*** Original change description ***

Defer TUN device creation until after the quic connection has been established.

IOS XR's XR container takes a few seconds for networking state to be sync'd from the host -- during which time any manipulations to the linux contaier's host networking stack breaks the synchronization process.

This has the nice side-benefit of re-initializing the TUN device each time a new quic connection is established, ensuring the local interface is in a good state on each reconnect. Previously, if the interfac...

***

PiperOrigin-RevId: 373010349