Automated g4 rollback of changelist 373010349.
*** Reason for rollback ***

This change no longer downs the interface when taking down the qbone tunnel. Instead, we close the fd we have for interacting with the TUN -- this has the effect of removing the interface when nopersist is set on the device.

*** Original change description ***

Automated g4 rollback of changelist 372620094.

*** Reason for rollback ***

Breaking Endor because we bring down the TUN Device when tearing down the bonnet in the init container.

*** Original change description ***

Defer TUN device creation until after the quic connection has been established.

IOS XR's XR container takes a few seconds for networking state to be sync'd from the host -- during which time any manipulations to the linux contaier's host networking stack breaks the synchronizatio...

***

PiperOrigin-RevId: 373210829