Retire server connection ID proactively on path validation failure instead of waiting for read/write other packets.
The alarm to retire connection IDs no longer on paths is scheduled at the
end of writing and reading packet. On path validation failure, there could
be no packet to write or read. In this case, the retirement alarm for the connection ID associated with the failed path needs to be proactively scheduled.

PiperOrigin-RevId: 376230134