Change GetServerConfig() to return const ref.
Change QuicTestClient::GetServerConfig() and SimpleQuicClient::GetServerConfig()
to return const ref instead of value in order to eliminate a copy.  This is only
used in one test file, where operator[] cannot be used any longer as it is not
constant.

Change QuicTestClient::GetServerConfig() to require that
QuicCryptoClientConfig::server_config_ to be set otherwise it crashes.  This is
safe since this method is only called by tests internally, and there are no
callers in Chrome.

Also remove unused QuicCryptoServerConfig::tag_value_map.

PiperOrigin-RevId: 378863306