Support logging ssl_handshake_processing_time for QUIC as well.
The purpose of this is to measure time taken by the GFE for the SSL handshake, not including time waiting for the client. I.e., it's just the aggregate time in the GFE process for the handshake.

I wasn't sure what the the best approach is, so please let me know if tracking this time in a different place would be better.

PiperOrigin-RevId: 379360130