Fix QUIC insecure randomness on Windows 7
These tests were disabled recently due to a rare crash on Windows 7. We tracked the crash down to the fact that taking the address of a static thread_local uint64_t[4] is apparently buggy on Windows 7. This change instead initializes that variable by assignment to sidestep the issue.

We've confirmed that this change fixes the Windows 7 Chrome crasher:
https://chromium-swarm.appspot.com/task?id=54e82650dc386a10

Note that this CL is not flag-protected because that would require two separate implementations of Xoshiro which would double the amount of thread_local storage required. That could cause issues on some platforms.

PiperOrigin-RevId: 386354470