Initialize SimpleRandom::buffer_offset_ in quic_test_utils
A few QUIC tests seemed like they might have been failing on Chromium's Win 7 bots because quic_test_utils's SimpleRandom::RandUint64 was being accessed out of bounds. [0] After the discussion in crbug.com/1228245#c23, this seems like a red herring: the initial stack trace was corrupted (the waterfall bots in question were running with a low symbol_level). The issue seems like it was actually in quic::DefaultRandom, not quic::test::SimpleRandom.

But, in any event, it seems like a safety improvement to initialize the buffer offset in SimpleRandom: this CL goes ahead and does so, even if it might not be necessary for addressing the crashes in question.

[0] https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket.appspot.com/8840889728099773792/+/u/net_unittests/Deterministic_failure:_QuicTimeWaitListManagerTest.SendStatelessResetInResponseToShortHeaders__status_CRASH_

PiperOrigin-RevId: 386564749