Add QUICHE_EXPORT macros in third_party/http2/adapter.
Also to FrameTypeToString() in third_party/spdy/core/spdy_protocol.h, which is
required for test_utils.cc to link in Chromium.

QUICHE_EXPORT_PRIVATE is used for definitions that are compiled in production
code but used from test code as well (otherwise linking the net_unittest
Chromium target fails).

QUICHE_NO_EXPORT is used in compilation units that are part of the net_unittest
build target, because these are not linked into any other target.  This macro is
defined empty both internally and in Chromium and is meant to supress the
presubmit error.

Also clean up dependencies using build_cleaner.

PiperOrigin-RevId: 386958565