Fix more signed-unsigned comparisons in QUICHE.
This is in preparation for landing some macro magic that would actually enforce
it in test comparison macros.

In some places I fixed the issue by adding 'u' to an integer literal.  In other
places I chose the path of least resistance and modified the build rule.  This
is okay for files that are not included in Chromium and for which there are no
immediate plans to include them.  In the long term we might change the Chromium
build rules for QUICHE in which case this all becomes moot.  Or if we don't, it
is obvious that there are violations in the given file, and they can be fixed
and the build rule change reverted before including them in the Chromium build.
The important part is to make sure currently included files cannot be modified
in a -Wsign-compare violating way, so as not to block QUICHE merge.

PiperOrigin-RevId: 389891302