Let client-side OgHttp2Session obey the peer's advertised SETTINGS_MAX_CONCURRENT_STREAMS.
This CL allows client-side OgHttp2Session to save the peer's advertised
SETTINGS_MAX_CONCURRENT_STREAMS value and use that value in SubmitRequest() to
determine whether to create a stream immediately for the request or store the
request for later. OgHttp2Session will create a stream immediately unless that
would violate SETTINGS_MAX_CONCURRENT_STREAMS; in this case instead, the
request is stored in a (newly added) PendingStreamState. Then when removing a
stream from the map in CloseStream(), OgHttp2Session will check if pending
streams can be created (compliant with SETTINGS_MAX_CONCURRENT_STREAMS).

PiperOrigin-RevId: 398554048