Do not preallocate memory in HttpDecoder for frame payload.
DATA, HEADERS, and unknown frame payload fragments are passed immediately to
Visitor without buffering.  CANCEL_PUSH and PUSH_PROMISE frames are not
processed.  GOAWAY and MAX_PUSH_ID frames have small payloads.  Only SETTINGS,
PRIORITY_UPDATE, and ACCEPT_CH frames may be buffered by HttpDecoder and are
allowed to have large payloads.  In most cases these frames are small and the
payload is sent in a single fragment, in which case HttpDecoder parses it
directly from the input without buffering.  Buffering only happens in the rest
of the cases, which clearly is not a hot path.  This CL changes HttpDecoder not
to preallocate memory for the payload buffer in these cases, and let std::string
deal with growing internally.

Also introduce kPayloadLengthLimit constant, and do not call MaxFrameLength()
for frames that are not buffered.

PiperOrigin-RevId: 399421301