Disambiguate MockVisitor
Chrome places multiple tests in one binary, and that does not work well when there are conflicting names that don't live in anonymous namespaces. This CL fixes the problem by adding both a belt and suspenders: it renames the visitor and adds an anonymous namespace to prevent folks from accidentally adding the problem back.

PiperOrigin-RevId: 400026052