Call visitor_.OnConnectionError() when failing to send the connection preface.
This CL brings the connection preface write failure handling in line with
queued frame write failure handling
(http://google3/third_party/http2/adapter/oghttp2_session.cc?l=332-334&rcl=399578585).

This CL also changes a static const member to a static constexpr member. For
some reason, this change was needed to be able to refer to the member outside
of the class (e.g., in a test).

PiperOrigin-RevId: 400037133