Automated g4 rollback of changelist 291452792.
*** Reason for rollback ***

QuicCryptoServerConfig is shared by all server connections, we are going to need ClientCertMode to be configured per-connection.

*** Original change description ***

Add a ServerProofVerifier and ClientCertMode to QuicCryptoServerConfig.

These two parameters allow configuration of server endpoints that request (or
require) client certificates.

This cl adds the settings, but implementation is left to a future change.

See go/quic-tls-client-certificates for details.

gfe-relnote: no behavior change (new settings are not used in gfe)

***

PiperOrigin-RevId: 401242157