Ignore user agent transport parameter
In gQUIC, we used to send the user agent string via the UAID tag on the CHLO. In IETF QUIC, the user agent is communicated using a regular HTTP header (and compressed using QPACK). While transitioning from gQUIC to IETF QUIC, we had added this custom Google-only transport parameter to facilitate migration. Now that we are getting production traffic from other browsers that do not send this transport parameter, we're confident that our infrastructure is correctly getting the user agent from the HTTP header and this transport parameter is therefore mostly redundant. We'll lose the ability to see the UAID before the first request but that's not much of a concern in a world where browsers are working towards freezing their user agent strings. Removing it also improves user privacy because it is sent in the clear and enables fingerprinting. Once this flag is deprecated, we will remove the transport parameter entirely in a follow-up CL.

Protected by FLAGS_quic_reloadable_flag_quic_ignore_user_agent_transport_parameter.

PiperOrigin-RevId: 402616442