Ignore key_update_not_yet_support transport parameter
When we first introduced IETF QUIC key update, we were communicating with older versions of our implementation that didn't support this feature. If we updated keys when communicating with that older software, the connection would fail. To avoid that, we had added a temporary key_update_not_yet_support transport parameter that would disable key updates. We stopped sending that transport parameter when we added support for key updates in Chrome m88. That version is no longer in any experiments, and also that version did not yet have IETF QUIC default-enabled, so we are guaranteed to not be communicating using IETF QUIC with this old software any more. It's therefore safe to remove this transport parameter as it no longer provides value. Once we deprecate the flag, we'll remove transport parameter entirely.

Protected by FLAGS_quic_reloadable_flag_quic_ignore_key_update_not_yet_supported.

PiperOrigin-RevId: 402628993