Remove QPACK fuzz corpora.
The fuzzing codelab makes the following recommendation: "Do not include in the
build rule corpus tests that were generated by the fuzzer running locally. This
is unnecessary, as the fuzzing infrastructure will automatically and
transparently preserve and reuse the corpus generated by fuzzer runs."  Since I
generated all files in the corpus by running the fuzzer locally, they should not
be checked in.

PiperOrigin-RevId: 404599498