Platformize AsciiUrlDecode
This function broke Chromium merge because of a different method signature for DecodeURLEscapeSequences. This CL allows us to override the method in Chromium. We've confirmed that this works in Chromium. This CL also fixes a bug where URI expansion was incorrectly leaving unknown variables instead of removing them. That issue was found by use the Chromium platform override in the corresponding merge CL.

PiperOrigin-RevId: 406948688