Refactor QUIC version downgrade prevention
This CL is the start of a series of CLs aimed at supporting draft-ietf-quic-version-negotiation. This CL refactors the Google QUIC version downgrade prevention transport parameter to better indicate that it is optional. This will make it easier to implement the IETF replacement and run both together during the transition period. This CL does not change any behavior, and is therefore not flag protected.

PiperOrigin-RevId: 408700446