Add testing to demonstrate nghttp2 and oghttp2 handling of data on a closed stream.
This CL adds an OgHttp2 and corresponding NgHttp2 unit test that exercises the
following scenario:
  1. Client sends request to open a stream.
  2. Client sends RST_STREAM on the stream it just opened.
  3. Client receives response HEADERS + DATA on the stream.

This scenario can validly arise when the server sends the response between
the receipt of Steps 1 and 2.

NgHttp2 does not fully process/deliver callbacks for the closed stream, whereas
OgHttp2 does. It is hypothesized that this difference in behavior is causing a
test failure in
Http2FloodMitigationTest.UpstreamRstStreamStormOnDownstreamCloseRegressionTest
as explained in http://b/201799377#comment8.

PiperOrigin-RevId: 410124780