Allow visitor_.OnFrameSent() to propagate a fatal send error, and do not retry sending on fatal send errors.
This CL makes OgHttp2Session::AfterFrameSent() return a bool to reflect an
error value from visitor_.OnFrameSent(), which is treated as a fatal connection
error and propagated to OgHttp2Session::Send(). This propagation is helpful for
compatibility with nghttp2 and Envoy.

For reference, the only (practically considered) error code nghttp2 returns for
failed sends is a fatal error:
https://nghttp2.org/documentation/nghttp2_session_send.html

This CL also makes a change to not let OgHttp2Session::Send() send data when it
has previously returned an error value. This is not consistent with nghttp2,
but it feels right and does not negatively affect current codec_impl_tests.

With this CL, codec_impl_tests Sponge (ProtocolErrorForTest now passes):
http://sponge2/279619fd-ba65-4ba9-ad96-367b4606cfc8

PiperOrigin-RevId: 416424214