Refactor OgHttp2Session::ProcessBytes().
This CL adds internal OgHttp2Session::ProcessBytesImpl() that returns an
absl::variant that contains either the number of bytes processed on success or
an error code on failure. Then ProcessBytes() calls ProcessBytesImpl() and
returns an overall int based on the absl::variant.

This CL is a pure refactoring CL with only one error code at the moment, but
the error codes will grow in a future CL to include values compatible with
nghttp2. This compatibility will be useful for some codec_impl_test tests.

PiperOrigin-RevId: 416872938