Add testing for content-length handling.
This CL adds oghttp2 and nghttp2 testing for content-length handling, including
scenarios where the actual data length is smaller/larger than the
content-length value and where the stream ends with a DATA vs. HEADERS frame.
Surprisingly and relevantly for Envoy integration testing, nghttp2 calls
OnInvalidFrame() when the stream ends with a HEADERS frame (but not DATA
frame). This CL shows content-length gaps that oghttp2 needs to fill.

PiperOrigin-RevId: 422594751